requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Session==0.5.0
scikit-learn==1.3.2
werkzeug==2.3.7
mysql-connector-python==8.1.0
PyPDF2==3.0.0
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
easyocr==1.7.0
sentence-transformers==2.2.2
numpy==1.24.3
requests==2.31.0
gunicorn==25.3.0
python-dotenv==1.0.1
google-generativeai==0.8.3
deep-translator==1.11.4
